Move sign in screen to be inside the advisor screen and rework navigation to place all screens in the same activity
