+++ android/app/src/main/AndroidManifest.xml
@@ -16,22 +16,13 @@
         android:theme="@style/Theme.GPAi"
         tools:targetApi="31">
         <activity
-            android:name=".SignInActivity"
+            android:name=".activity.HomeActivity"
             android:exported="true"
-            android:label="@string/title_activity_sign_in"
             android:theme="@style/Theme.GPAi">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".onboarding.OnboardingActivity"
-            android:theme="@style/Theme.GPAi">
-        </activity>
-        <activity
-            android:name=".home.HomeActivity"
-            android:theme="@style/Theme.GPAi">
         </activity>
     </application>
